refactor: :art: move format and logical validators to templates
