+++ src/main/java/mantenimiento/codecounter/templates/LogicalValidator.java
@@ -1,10 +1,9 @@
-package mantenimiento.codecounter.interfaces;
+package mantenimiento.codecounter.templates;
 
 /*
  * Interfaz que define el comportamiento de una fábrica de validadores lógicos
  */
-public interface LogicalValidatorFactory {
-
+public interface LogicalValidator {
   /**
    * Valida si la línea de código corresponde a una declaración de tipo
    *
