+++ src/main/java/mantenimiento/codecounter/validators/logicalValidators/TypeDeclarationValidator.java
@@ -3,13 +3,13 @@ package mantenimiento.codecounter.validators.logicalValidators;
 import static mantenimiento.codecounter.constants.JavaRegextConstants.ACCESS_MODIFIERS_REGEX;
 import static mantenimiento.codecounter.constants.JavaRegextConstants.TYPE_KEYS;
 
-import mantenimiento.codecounter.interfaces.LogicalValidatorFactory;
+import mantenimiento.codecounter.templates.LogicalValidator;
 
 /**
  * Clase que verifica si existe una declaración de tipo (class, interface, enum) para poder
  * contabilizarlo como línea lógica
  */
-public class TypeDeclarationValidator implements LogicalValidatorFactory {
+public class TypeDeclarationValidator implements LogicalValidator {
 
   private static final String TYPE_DECLARATION =
       "^(\\s*" + ACCESS_MODIFIERS_REGEX + ".*\\s*" + TYPE_KEYS + ".*)";
